Moved IFilesWords from IPerister to a member variables
