+++ src/myoddweb.desktopsearch.service/Persisters/SqlitePersisterWords.cs
@@ -89,6 +89,71 @@ namespace myoddweb.desktopsearch.service.Persisters
         connectionFactory, token).ConfigureAwait(false);
     }
 
+    /// <inheritdoc />
+    public async Task<List<long>> GetWordIdsAsync(Words words, IConnectionFactory connectionFactory, CancellationToken token, bool createIfNotFound)
+    {
+      // do we have anything to even look for?
+      if (words.Count == 0)
+      {
+        return new List<long>();
+      }
+
+      try
+      {
+        // we do not check for the token as the underlying functions will throw if needed. 
+        // look for the word, add it if needed.
+        var sql = $"SELECT id FROM {TableWords} WHERE word=@word";
+        using (var cmd = connectionFactory.CreateCommand(sql))
+        {
+          var pWord = cmd.CreateParameter();
+          pWord.DbType = DbType.String;
+          pWord.ParameterName = "@word";
+          cmd.Parameters.Add(pWord);
+
+          var ids = new List<long>(words.Count);
+
+          // we use a list to allow duplicates.
+          var wordsToAdd = new List<Word>();
+          foreach (var word in words)
+          {
+            pWord.Value = word.Value;
+            var value = await connectionFactory.ExecuteReadOneAsync(cmd, token).ConfigureAwait(false);
+            if (null != value && value != DBNull.Value)
+            {
+              // get the path id.
+              ids.Add((long)value);
+              continue;
+            }
+
+            if (!createIfNotFound)
+            {
+              // we could not find it and we do not wish to go further.
+              ids.Add(-1);
+              continue;
+            }
+
+            // add this word to our list
+            wordsToAdd.Add(word);
+          }
+
+          // finally we need to add all the words that were not found.
+          ids.AddRange(await AddOrUpdateWordsAsync(new Words(wordsToAdd.ToArray()), connectionFactory, token).ConfigureAwait(false));
+
+          // return all the ids we added, (or not).
+          return ids;
+        }
+      }
+      catch (OperationCanceledException)
+      {
+        _logger.Warning("Received cancellation request - Get multiple word ids.");
+        throw;
+      }
+      catch (Exception ex)
+      {
+        _logger.Exception(ex);
+        throw;
+      }
+    }
     #region Private word functions
     /// <summary>
     /// Given a list of words, re-create the ones that we need to insert.
